src/domain/forum/application/use-cases/fetch-question-comments.spec.ts
import { UniqueEntityID } from '@/core/entities/unique-entity';
import { makeQuestionComment } from 'test/factories/make-question-comment';
import { InMemoryQuestionCommentsRepository } from 'test/repositories/in-memory-question-comments-repository';
import { expect, it } from 'vitest';
import { FetchQuestionCommentsUseCase } from './fetch-question-comments';

let inMemoryQuestionCommentRepository: InMemoryQuestionCommentsRepository
let sut: FetchQuestionCommentsUseCase

describe('Fetch Question Comments', () => {
  beforeEach(() => {
    inMemoryQuestionCommentRepository = new InMemoryQuestionCommentsRepository();
    sut = new FetchQuestionCommentsUseCase(inMemoryQuestionCommentRepository)
  })

  it('should be able to fetch question comments', async () => {
    await inMemoryQuestionCommentRepository.create(makeQuestionComment({ questionId: new UniqueEntityID('question-1') }))
    await inMemoryQuestionCommentRepository.create(makeQuestionComment({ questionId: new UniqueEntityID('question-1') }))
    await inMemoryQuestionCommentRepository.create(makeQuestionComment({ questionId: new UniqueEntityID('question-1') }))

    const result = await sut.execute({
      questionId: 'question-1',
      page: 1,
    });

    expect(result.isRight()).toBe(true);
    const questionComments = result.value?.questionComments;
    expect(questionComments).toHaveLength(3)
  });

  it('should be able to fetch paginated question comments', async () => {
    for (let i = 1; i <= 28; i++) {
      await inMemoryQuestionCommentRepository.create(makeQuestionComment({ questionId: new UniqueEntityID('question-1') }))
    }
    
    const result = await sut.execute({
      questionId: 'question-1',
      page: 2,
    });
  
    expect(result.isRight()).toBe(true);
    const questionComments = result.value?.questionComments;
    expect(questionComments).toHaveLength(8)
  })
})

